Time: 102 ms (86.34%), Space: 85.9 MB (40.06%) - LeetHub

+++ 1770-maximum-score-from-performing-multiplication-operations/1770-maximum-score-from-performing-multiplication-operations.cpp
@@ -13,14 +13,14 @@ public:
     int maximumScore(vector<int>& nums, vector<int>& multipliers) {
         int m = multipliers.size();
         int n = nums.size();
-        vector<vector<long>> dp(m+1, vector<long>(m+1, 0));
+        vector<vector<int>> dp(m+1, vector<int>(m+1, 0));
         for(int i=m-1; i>=0; i--)
             for(int left=i; left>=0; left--){
                 int mult = multipliers[i];
                 int right = n-1 - (i-left);
                 dp[i][left] = max(
-                    (long)nums[left] * mult + dp[i+1][left+1],
-                    (long)nums[right] *mult + dp[i+1][left]
+                    nums[left] * mult + dp[i+1][left+1],
+                    nums[right] *mult + dp[i+1][left]
                 );
             }
